backend_langgraph/requirements.txt
langchain
langchain_core
langchain_anthropic
langchain_openai
tavily_python
langchain_community
pyairbnb==1.0.9
langgraph
geopy
pydantic[email]
python-dotenv
fastapi
uvicorn
jinja2
pydantic
sendgrid
tavily_python
langchain_groq
fast-flights
PyPDF2
beautifulsoup4
googlemaps
docling
arxiv
needle-python
openai
reportlab
aiohttp
ipython
primp==0.11.0
python-docx 
docx2pdf
pdfkit
langfuse
langchain-tavily
fastapi>=0.115.12
langchain>=0.3.25
langchain-core>=0.3.58
langchain-openai>=0.3.16
langfuse==3.0.3
langgraph>=0.4.1
langgraph-checkpoint-postgres>=2.0.19
passlib[bcrypt]>=1.7.4
psycopg2-binary>=2.9.10
pydantic[email]>=2.11.1
pydantic-settings>=2.8.1
python-dotenv>=1.1.0
python-jose[cryptography]>=3.4.0
python-multipart>=0.0.20
sqlmodel>=0.0.24
structlog>=25.2.0
supabase>=2.15.0
uvicorn>=0.34.0
bcrypt>=4.3.0
slowapi>=0.1.9
email-validator>=2.2.0
prometheus-client>=0.19.0
starlette-prometheus>=0.7.0
asgiref>=3.8.1
duckduckgo-search>=3.9.0
langchain-community>=0.3.20
tqdm>=4.67.1
colorama>=0.4.6
langserve
langchain-mcp-adapters
elevenlabs
langgraph-cli[inmem]>=0.3.8
livekit>=1.0.12
livekit-agents[cartesia,deepgram,openai,silero,turn-detector]~=1.2
livekit-plugins-cartesia>=1.2.6
livekit-plugins-deepgram>=1.2.6
livekit-plugins-hume>=1.2.6
livekit-plugins-langchain~=1.1
livekit-plugins-noise-cancellation~=0.2
livekit-plugins-openai>=1.2.6
livekit-plugins-silero>=1.2.6
livekit-plugins-turn-detector>=1.2.6
livekit-plugins-elevenlabs>=1.2.6